refactored units conversion into separate file
